LayoutTests/fast/frames/seamless/seamless-inline-expected.txt

Test that inline seamless iframes 'shrink-wrap' their contents like inline-blocks do.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.getComputedStyle(iframe1).display is "inline-block"
PASS window.getComputedStyle(iframe2).display is "inline-block"
PASS window.getComputedStyle(iframe1).width is "150px"
FAIL window.getComputedStyle(iframe1).height should be 50px. Was 104px.
FAIL window.getComputedStyle(parent1).height should be 50px. Was 108px.
PASS window.getComputedStyle(iframe2).width is "100px"
FAIL window.getComputedStyle(iframe2).height should be 100px. Was 104px.
FAIL window.getComputedStyle(parent2).height should be 150px. Was 162px.
